refactor: product price 분리
